examples/layout-tests/serviceworker/current-on-load.html

<!DOCTYPE html>
<title>Service Worker: Current on load</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/testutils.js"></script>
<body>
<script>
var t = async_test('current is set for a controlled document');
t.step(function() {
    var scope = 'resources/blank.html'
    navigator.serviceWorker.unregister(scope).then(doTest, doTest);

    function doTest() {
        navigator.serviceWorker.register(
            'resources/worker-no-op.js', {scope: scope}
        ).then(onRegister, t.step_func(function(reason) {
            assert_unreached('Registration should succeed, but failed: ' + reason.name);
        }));
    }

    function onRegister(worker) {
        // FIXME: Remove this check after crbug.com/365252 is fixed.
        if (worker.state == 'active')
            onActive();
        worker.addEventListener('statechange', t.step_func(function(event) {
            if (event.target.state == 'active')
                onActive();
        }));
    }

    function onActive() {
        withIframe(scope, t.step_func(function(frame) {
            var w = frame.contentWindow;
            assert_true(w.navigator.serviceWorker.current instanceof ServiceWorker,
                        'current should be a ServiceWorker object');
            t.done();
        }));
    }
});
</script>
</body>
